requirements.txt
asgiref==3.5.2
black==22.12.0
cfgv==3.3.1
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.1.3
django-cors-headers==4.1.0
django-import-export==3.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
docformatter==1.5.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
filelock==3.8.2
flake8==6.0.0
identify==2.5.9
inflection==0.5.1
isort==5.11.1
MarkupPy==1.14
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pre-commit==2.20.0
pre-commit-hooks==4.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
untokenize==0.1.1
uritemplate==4.1.1
virtualenv==20.17.1
xlrd==2.0.1
xlwt==1.3.0
